Fixed warnings about signed/unsigned comparison.

+++ src/template/libxsmm_dnn_convolution_winograd_backward_nhwc_custom_input_trans_alpha6_avx512.tpl.c
@@ -33,8 +33,8 @@ const int total_tiles = handle->cwino_bwd.itiles*handle->cwino_bwd.jtiles;
 LIBXSMM_VLA_DECL(4, const float, input, inp, handle->ofwp, handle->blocksofm, TDVLEN);
 LIBXSMM_VLA_DECL(6, float, output, tinp, ALPHA, handle->cwino_bwd.bimg, total_tiles, handle->blocksofm, TDVLEN);
 __m512 I[ALPHA];
-int ti, tj;
-int i, j;
+unsigned int ti, tj;
+unsigned int i, j;
 int xdim, ydim;
 const int l_pad = (handle->desc.W - handle->ofw)/2 + 1;
 const int t_pad = (handle->desc.H - handle->ofh)/2 + 1;
@@ -43,8 +43,8 @@ __m512 t0, t1, t2, t3, t4, t5;
 
 for (tj = 0; tj < handle->cwino_bwd.jtiles; tj++) {
   for (ti = 0; ti < handle->cwino_bwd.itiles; ti++) { /* for each tile */
-    if (ti*((ALPHA)-2) >= l_pad && ti*((ALPHA)-2) + (ALPHA) <= (handle->ofw + l_pad) &&
-        tj*((ALPHA)-2) >= t_pad && tj*((ALPHA)-2) + (ALPHA) <= (handle->ofh + t_pad)) { /* common case */
+    if (ti*((ALPHA)-2) >= (unsigned int)l_pad && ti*((ALPHA)-2) + (ALPHA) <= (unsigned int)(handle->ofw + l_pad) &&
+        tj*((ALPHA)-2) >= (unsigned int)t_pad && tj*((ALPHA)-2) + (ALPHA) <= (unsigned int)(handle->ofh + t_pad)) { /* common case */
 
       /* left multiplication */
       /* this unrolling didn't help performance much so we may want to remove later if code size becomes an issue */
